+++ changelog.md
@@ -17,6 +17,7 @@
 - 29-Oct-2021 Port VU meters to v2, including new backlights
 - 03-Nov-2021 Make DO-1xx devices more usable, improve connection labelling, add shuffle and copy/paste
 - 05-Nov-2021 Add visuals to TD-xxx, DO-1xx and WM-102 devices in the module browser
+- 05-Nov-2021 Add Device copy/paste and undo to AO-1xx devices
 
 ## v1.1.10
 - 10-Dec-2020 Update images in manual